requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg==1.21.7
djangorestframework
django-filter==24.2
django-modeltranslation
django-redis
cryptography==41.0.7
django-recaptcha==3.0.0
nplusone==1.0.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
dj-database-url==2.2.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
djangorestframework-simplejwt==5.5.0
requests==2.32.3
django-celery-beat==2.8.1
moviepy==2.2.1
channels==4.2.2
daphne==4.2.0
channels_redis==4.2.1
weasyprint==65.1